refactor: 난이도 매핑 로직 및 상수 리팩토링
- mapDifficulty 함수를 `difficulty.util.ts`로 이동하여 재사용성 향상
- 기존 난이도 매핑 및 범위 계산 함수들을 통합/삭제하고 공통 유틸 함수 사용
- quiz.constants.ts와 question.transformer.ts 등 관련 파일에서 중복 코드 제거
- 문제 은행 서비스에서 난이도 필터 관련 함수 간소화 및 가독성 개선
