+++ packages/backend/src/game/round-progression.service.ts
@@ -4,11 +4,12 @@ import { RoundTimer } from './round-timer';
 import { QuizService } from '../quiz/quiz.service';
 import { GameSessionManager } from './game-session-manager';
 import { MatchPersistenceService } from './match-persistence.service';
-import { Difficulty, getValidQuestionType, ROUND_DURATIONS } from './round-timer.constants';
+import { getValidQuestionType, ROUND_DURATIONS } from './round-timer.constants';
 import { SPEED_BONUS } from '../quiz/quiz.constants';
 import { transformQuestionForClient } from './transformers/question.transformer';
 import { FinalResult } from './interfaces/game.interfaces';
 import { RoundResult } from '../quiz/quiz.types';
+import { mapDifficulty } from '../common/utils/difficulty.util';
 
 @Injectable()
 export class RoundProgressionService {
@@ -92,17 +93,7 @@ export class RoundProgressionService {
       });
 
       // 난이도 매핑 (DB의 숫자 난이도 -> 문자열)
-      const difficultyNum = question.difficulty || 3;
-      let difficulty: Difficulty;
-
-      if (difficultyNum <= 2) {
-        difficulty = 'easy';
-      } else if (difficultyNum === 3) {
-        difficulty = 'medium';
-      } else {
-        difficulty = 'hard';
-      }
-
+      const difficulty = mapDifficulty(question.difficulty);
       const questionType = getValidQuestionType(question.questionType);
       const questionDuration = ROUND_DURATIONS.QUESTION[questionType][difficulty];
