+++ packages/backend/src/quiz/quiz.constants.ts
@@ -1,5 +1,7 @@
 import { Difficulty } from './quiz.types';
 
+export { mapDifficulty } from '../common/utils/difficulty.util';
+
 export const SCORE_MAP: Record<Difficulty, number> = {
   easy: 10,
   medium: 20,
@@ -7,26 +9,6 @@ export const SCORE_MAP: Record<Difficulty, number> = {
 };
 
 export const SPEED_BONUS = 5;
-
-/**
- * 숫자 난이도를 문자열 난이도로 매핑
- * 1-2: easy, 3: medium, 4-5: hard
- */
-export function mapDifficulty(numDifficulty: number | null): Difficulty {
-  if (!numDifficulty) {
-    return 'medium';
-  }
-
-  if (numDifficulty <= 2) {
-    return 'easy';
-  }
-
-  if (numDifficulty === 3) {
-    return 'medium';
-  }
-
-  return 'hard';
-}
 
 /**
  * Quiz 서비스 상수 정의
